Refactor restaurant dashboard views: improve booking, menu, post, and review listings with better structure and no data found messages.

+++ app/Views/restaurant_dashboard/post.php
@@ -2,11 +2,11 @@
 
 <div class="post-container">
     <div class="top">
-            <h1>Post List</h1><span></span>
+        <h1>Post List</h1><span></span>
 
-            <div class="action-buttons">
-                <a class="add-btn" href="?page=post&action=add">Add Post</a>
-            </div>
+        <div class="action-buttons">
+            <a class="add-btn" href="?page=post&action=add">Add Post</a>
+        </div>
     </div>
 
     <table>
@@ -21,21 +21,26 @@
         </thead>
 
         <tbody>
+            <?php if (!empty($posts) && is_array($posts)): ?>
+                <?php foreach ($posts as $post) : ?>
+                    <tr>
+                        <td><?= $post['Title'] ?></td>
+                        <td><?= htmlspecialchars($post['Description']) ?></td>
+                        <td> <?= date('d-m-Y', strtotime($post['Date'])) ?></td>
+                        <td><img src="<?= $post['ImgPath'] ?>" class="post-img"></td>
+                        <td class="action-buttons">
+                            <button class="edit-btn"><a href="?page=post&action=edit&id=<?= $post['PostID'] ?>">Edit</a></button>
+                            <button class="delete-btn"> <a href="?page=post&action=delete&id=<?= $post['PostID'] ?>">Delete</button>
+                        </td>
+                    </tr>
 
-        <?php foreach ($posts as $post) : ?>
-            <tr>
-                <td><?= $post['Title'] ?></td>
-                <td><?= htmlspecialchars($post['Description']) ?></td>
-                <td> <?= date('d-m-Y', strtotime($post['Date'])) ?></td>
-                <td><img src="<?= $post['ImgPath'] ?>" class="post-img"></td>
-                <td class="action-buttons">
-                    <button class="edit-btn"><a href="?page=post&action=edit&id=<?= $post['PostID'] ?>">Edit</a></button>
-                    <button class="delete-btn">  <a href="?page=post&action=delete&id=<?= $post['PostID'] ?>">Delete</button>
-                </td>
-            </tr>
-             
-          <?php endforeach; ?> 
- 
+                <?php endforeach; ?>
+            <?php else: ?>
+                <tr>
+                    <td colspan="5" style="text-align: center;">No posts found.</td>
+                </tr>
+            <?php endif; ?>
+
         </tbody>
     </table>
 </div>
